feat: added (test) Stripe payment API to checkout->Order

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const mongoose = require('mongoose')
+const fileHelper = require('../util/file')
 
 const { validationResult } = require('express-validator/check');
 
@@ -155,6 +156,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl)
         product.imageUrl = image.path;
       }
       return product.save().then(result => {
@@ -188,12 +190,20 @@ exports.getProducts = (req, res, next) => {
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => console.log(err));
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
+  Product.findById(prodId).then(product => {
+    if (!product) {
+      return next(new Error('Product not found.'))
+    }
+    fileHelper.deleteFile(product.imageUrl)
+    return Product.deleteOne({ _id: prodId, userId: req.user._id })
+  })
+  .then(() => {
+    console.log('DESTROYED PRODUCT');
+    res.status(200).json({message: 'Success!'})
+  })
+  .catch(err => {
+    res.status(500).json({message: 'Deleting product failed.'})
+  });
 };
